Tests fixed.

+++ src/main/scala/org/ergoplatform/explorer/grabber/Constants.scala
@@ -3,6 +3,6 @@ package org.ergoplatform.explorer.grabber
 object Constants {
   val TestnetPrefix: Byte = 0x10
   val MainnetPrefix: Byte = 0x00
-  val PreGenesisHeight: Int = 0
-  val GenesisHeight: Int = PreGenesisHeight + 1
+  val PreGenesisHeight: Long = 0L
+  val GenesisHeight: Long = PreGenesisHeight + 1
 }
